requirements.txt
APScheduler==3.11.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
Flask==3.0.3
Flask-Cors==5.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pymongo==4.10.1
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
Werkzeug==3.0.4
